Add ability to interact with LiveSplit One sources
OBS sources provide an “Interact” window, which allows mouse and
keyboard interaction with the individual source.  The OBS LiveSplit
One plugin’s Interact window only allows for scrolling up and down
through splits, but this can be useful with layouts that do not
display all splits in the run.

This patch implements a new action for triggering the Interact window
of any LiveSplit One source that OBS knows about.  Unlike the hotkey
actions, this action supports per-action source configuration, for
cases where multiple sources are used in-tandem in an OBS scene.